libwasi: make integer truncation explicit

+++ libwasi/wasi_abi_fd.c
@@ -208,7 +208,8 @@ retry:
                 ret = EOVERFLOW;
                 goto fail;
         }
-        host_ret = wasi_copyout_result_i32(ctx, wasi_memory(wasi), retp, n);
+        host_ret = wasi_copyout_result_i32(ctx, wasi_memory(wasi), retp,
+                                           (uint32_t)n);
         ret = 0;
 fail:
         wasi_fdinfo_release(wasi, fdinfo);
@@ -265,7 +266,8 @@ retry:
                 ret = EOVERFLOW;
                 goto fail;
         }
-        host_ret = wasi_copyout_result_i32(ctx, wasi_memory(wasi), retp, n);
+        host_ret = wasi_copyout_result_i32(ctx, wasi_memory(wasi), retp,
+                                           (uint32_t)n);
         ret = 0;
 fail:
         wasi_fdinfo_release(wasi, fdinfo);
@@ -338,7 +340,8 @@ tty_hack:
                 ret = EOVERFLOW;
                 goto fail;
         }
-        host_ret = wasi_copyout_result_i32(ctx, wasi_memory(wasi), retp, n);
+        host_ret = wasi_copyout_result_i32(ctx, wasi_memory(wasi), retp,
+                                           (uint32_t)n);
         ret = 0;
 fail:
         wasi_fdinfo_release(wasi, fdinfo);
@@ -395,7 +398,8 @@ retry:
                 ret = EOVERFLOW;
                 goto fail;
         }
-        host_ret = wasi_copyout_result_i32(ctx, wasi_memory(wasi), retp, n);
+        host_ret = wasi_copyout_result_i32(ctx, wasi_memory(wasi), retp,
+                                           (uint32_t)n);
         ret = 0;
 fail:
         wasi_fdinfo_release(wasi, fdinfo);
